+++ services/fitbit/FitbitManager.js
@@ -1,64 +1,24 @@
-const DateTimeUtils = require("./DateTimeUtils");
-const { UserTokenNotFoundError } = require("../../source/Errors");
 const Logger = require("../../source/Loggers");
-const ApiManager = require("./api/ApiManager");
 const DBManager = require("./db/DBManager");
-const FitbitHelper = require("./FitbitHelper");
 const Config = require("./Config");
 const Scheduler = require("./Scheduler");
 const RateLimit = require("./api/RateLimit");
+const RequestProcessor = require("./RequestProcessor");
 
 class FitbitManager {
 	static async registerUser(authorizationCode, userId) {
 		try {
-			const { access_token: accessToken,
-				expires_in: expiresIn,
-				refresh_token: refreshToken,
-				scope, user_id: fitbitUserId
-			} = await ApiManager.authorizeUser(authorizationCode);
-
-			let user = await DBManager.getUser(userId);
-			let tokenLastUpdated = DateTimeUtils.getCurrentDateTime();
-			let expirationDate = DateTimeUtils.getExpirationDateTime(expiresIn);
-
-			if (user) {
-				DBManager.updateUser(user, fitbitUserId, tokenLastUpdated,
-					accessToken, expirationDate, refreshToken, scope);
-			} else {
-				// Membership creation date in the fitbit account. This date is fetched once during the creation of the database entry.
-				let memberSince = await FitbitHelper.getMemberSince(accessToken, fitbitUserId);
-
-				DBManager.insertUser(userId, fitbitUserId, memberSince, tokenLastUpdated,
-					accessToken, expirationDate, refreshToken, scope);
-			}
+			await RequestProcessor.processRegistration(authorizationCode, userId);
 		} catch (error) {
 			Logger.error(error);
 			throw error;
 		}
 	}
 
-	static async refreshUserToken(userId) {
+	static async getUpdatedRefreshToken(userId) {
 		try {
-			let user = await DBManager.getUser(userId);
-
-			if (user) {
-				// Get the refresh token through the Fitbit API Wrapper
-				const { access_token: accessToken,
-					expires_in: expiresIn,
-					refresh_token: refreshToken,
-					user_id: fitbitUserId
-				} = await ApiManager.refreshToken(user.refresh_token);
-				let tokenLastUpdated = DateTimeUtils.getCurrentDateTime();
-				let expirationDate = DateTimeUtils.getCurrentDateTime(expiresIn);
-
-				// Insert updated user data into database
-				DBManager.updateAndFetchUser(user, fitbitUserId, tokenLastUpdated, accessToken,
-					expirationDate, refreshToken);
-
-				return user;
-			} else {
-				throw new UserTokenNotFoundError;
-			}
+			let user = await RequestProcessor.processRefreshToken(userId)
+			return user;
 		} catch (error) {
 			Logger.error(error);
 			throw error;
@@ -108,8 +68,7 @@ class FitbitManager {
 	static async pollUserData(userId) {
 		let tokenData;
 		try {
-			tokenData = await this.refreshUserToken(userId);
-			// console.log(tokenData)
+			tokenData = await this.getUpdatedRefreshToken(userId);
 		}
 		catch (error) {
 			Logger.error("Could not refresh access token for user " + userId + ": ", error, JSON.stringify(error));
@@ -122,17 +81,17 @@ class FitbitManager {
 			Logger.error("Error while processing Activities data for user " + userId + ":", error, JSON.stringify(error));
 		}
 
-		// try {
-		// 	await this.processTimeSeriesData(userId, tokenData.access_token, tokenData.fitbit_user_id);
-		// } catch (error) {
-		// 	Logger.error("Error while processing Time Series data for user " + userId + ":", error, JSON.stringify(error));
-		// }
+		try {
+			await this.processTimeSeriesData(userId, tokenData.access_token, tokenData.fitbit_user_id);
+		} catch (error) {
+			Logger.error("Error while processing Time Series data for user " + userId + ":", error, JSON.stringify(error));
+		}
 		
-		// try {
-		// 	await this.processIntradayData(userId, tokenData.access_token, tokenData.fitbit_user_id);
-		// } catch (error) {
-		// 	Logger.error("Error while processing Intraday data for user " + userId + ":", error, JSON.stringify(error));
-		// }
+		try {
+			await this.processIntradayData(userId, tokenData.access_token, tokenData.fitbit_user_id);
+		} catch (error) {
+			Logger.error("Error while processing Intraday data for user " + userId + ":", error, JSON.stringify(error));
+		}
 
 		console.log(RateLimit.processedUsers)
 		console.log("Resetting the number of request processed to 0!");
@@ -141,76 +100,37 @@ class FitbitManager {
 
 	static async processActivitiesData(userId, accessToken, fitbitUserId) {
 		try {
-			await this.processSingleCall(userId, accessToken, fitbitUserId, Config.requestType.activityStatistics);
-			await this.processSingleCall(userId, accessToken, fitbitUserId, Config.requestType.activityGoals);
+			await RequestProcessor.processSingleRequest(userId, accessToken, fitbitUserId, Config.requestType.activityStatistics);
+			await RequestProcessor.processSingleRequest(userId, accessToken, fitbitUserId, Config.requestType.activityGoals);
 			// The endpoint for Activity Log List is correct but the request failed with status code 400.
 			// Here is the url for the endpoint: https://dev.fitbit.com/build/reference/web-api/activity/get-activity-log-list/
-			await this.processSingleCall(userId, accessToken, fitbitUserId, Config.requestType.activityLogList);
-			await this.processSingleCall(userId, accessToken, fitbitUserId, Config.requestType.favoriteActivities);
-			await this.processSingleCall(userId, accessToken, fitbitUserId, Config.requestType.frequentActivities);
-			await this.processSingleCall(userId, accessToken, fitbitUserId, Config.requestType.recentActivities);
+			// await RequestProcessor.processSingleRequest(userId, accessToken, fitbitUserId, Config.requestType.activityLogList);
+			await RequestProcessor.processSingleRequest(userId, accessToken, fitbitUserId, Config.requestType.favoriteActivities);
+			await RequestProcessor.processSingleRequest(userId, accessToken, fitbitUserId, Config.requestType.frequentActivities);
+			await RequestProcessor.processSingleRequest(userId, accessToken, fitbitUserId, Config.requestType.recentActivities);
 		} catch (error) {
 			Logger.error("Error while processing Summary or Details data of activities for user " + userId + ":", error, JSON.stringify(error));
 		}
 
 		try {
-			await this.processRequestByDate(userId, accessToken, fitbitUserId, Config.requestType.dailyActivitySummary);
+			await RequestProcessor.processRequestByDate(userId, accessToken, fitbitUserId, Config.requestType.dailyActivitySummary);
 		} catch (error) {
 			Logger.error("Error while processing Summary or Details data of activities by date for user " + userId + ":", error, JSON.stringify(error));
 		}
 	}
 
-	static async processSingleRequest(userId, accessToken, fitbitUserId, requestType) {
-		console.log("Processing "+ requestType);
-		console.log("Number of request remainig in processing "+ requestType + " : 1");
-		if(RateLimit.isLimitExceeded()) {
-			console.log("Request limit is exceeded!");
-			RateLimit.setProcessedStatus(userId, false);
-			return;
-		}
-
-		let response = await ApiManager.getSummary(accessToken, fitbitUserId, requestType);
-		await DBManager.storeSummaryData(userId, requestType, response);
-
-		RateLimit.requestProcessed();
-		console.log("Total " + RateLimit.numberOfRequestProcessed + " Request processed successfully!");
-	}
-
-	static async processRequestByDate(userId, accessToken, fitbitUserId, requestType) {
-		console.log("Processing "+ requestType);
-		let startTimestamp = await FitbitHelper.getLastPolledTimestamp(userId, requestType);
-		let endTimestamp = await FitbitHelper.getLastSyncedTimestamp(accessToken, fitbitUserId);
-		const ranges = FitbitHelper.getTimeRanges(startTimestamp, endTimestamp);
-
-		console.log("RateLimit: "+RateLimit.totalQuota)
-		console.log("Refill: "+RateLimit.remainingSecondsUntilRefill)
-		console.log("Number of request remainig in processing "+ requestType + " : " + ranges.length);
-		for (const range of ranges) {
-			if(RateLimit.isLimitExceeded()) {
-				console.log("Request limit is exceeded!");
-				RateLimit.setProcessedStatus(userId, false);
-				break;
-			}
-			let response = await ApiManager.getSummaryByDate(accessToken, fitbitUserId, requestType, range.date);
-			await DBManager.storeSummaryData(userId, requestType, range, response);
-
-			RateLimit.requestProcessed();
-		}
-		console.log("Total " + RateLimit.numberOfRequestProcessed + " Request processed successfully!");
-	}
-
 	static async processTimeSeriesData(userId, accessToken, fitbitUserId) {
 		try {
-			await this.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.foodLogsCalories, 1095);
-			await this.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.foodLogsWater, 1095);
+			await RequestProcessor.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.foodLogsCalories, 1095);
+			await RequestProcessor.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.foodLogsWater, 1095);
 		} catch (error) {
 			Logger.error("Error while processing food log data for user " + userId + ":", error, JSON.stringify(error));
 		}
 
 		try {
-			await this.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.bodyBmi, 1095);
-			await this.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.bodyFat, 1095);
-			await this.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.bodyWeight, 1095);
+			await RequestProcessor.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.bodyBmi, 1095);
+			await RequestProcessor.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.bodyFat, 1095);
+			await RequestProcessor.processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, Config.requestType.bodyWeight, 1095);
 		} catch (error) {
 			Logger.error("Error while processing body bmi, fat and weight data for user " + userId + ":", error, JSON.stringify(error));
 		}
@@ -218,103 +138,34 @@ class FitbitManager {
 
 	static async processIntradayData(userId, accessToken, fitbitUserId) {
 		try {
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.heart, Config.detailLevel.oneSecond);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.heart, Config.detailLevel.oneSecond);
 		} catch (error) {
 			Logger.error("Error while processing heart rate data for user " + userId + ":", error, JSON.stringify(error));
 		}
 
 		try {
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.activeZoneMinutes, Config.detailLevel.oneMinute);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.activeZoneMinutes, Config.detailLevel.oneMinute);
 		} catch (error) {
 			Logger.error("Error while processing active zone minutes data for user " + userId + ":", error, JSON.stringify(error));
 		}
 
 		try {
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.calories, Config.detailLevel.oneMinute);
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.distance, Config.detailLevel.oneMinute);
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.elevation, Config.detailLevel.oneMinute);
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.floors, Config.detailLevel.oneMinute);
-			await this.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.steps, Config.detailLevel.oneMinute);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.calories, Config.detailLevel.oneMinute);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.distance, Config.detailLevel.oneMinute);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.elevation, Config.detailLevel.oneMinute);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.floors, Config.detailLevel.oneMinute);
+			await RequestProcessor.processIntraday(userId, accessToken, fitbitUserId, Config.requestType.steps, Config.detailLevel.oneMinute);
 		} catch (error) {
 			Logger.error("Error while processing activities data for user " + userId + ":", error, JSON.stringify(error));
 		}
 
 		try {
-			await this.processIntradayByInterval(userId, accessToken, fitbitUserId, Config.requestType.breathingRate, 30);
-			await this.processIntradayByInterval(userId, accessToken, fitbitUserId, Config.requestType.heartRateVariability, 30);
-			await this.processIntradayByInterval(userId, accessToken, fitbitUserId, Config.requestType.spO2, 30);
+			await RequestProcessor.processIntradayByInterval(userId, accessToken, fitbitUserId, Config.requestType.breathingRate, 30);
+			await RequestProcessor.processIntradayByInterval(userId, accessToken, fitbitUserId, Config.requestType.heartRateVariability, 30);
+			await RequestProcessor.processIntradayByInterval(userId, accessToken, fitbitUserId, Config.requestType.spO2, 30);
 		} catch (error) {
 			Logger.error("Error while processing intraday data by interval for user " + userId + ":", error, JSON.stringify(error));
 		}
-	}
-
-	static async processIntraday(userId, accessToken, fitbitUserId, requestType, detailLevel) {
-		console.log("Processing "+ requestType);
-		let startTimestamp = await FitbitHelper.getLastPolledTimestamp(userId, requestType);
-		let endTimestamp = await FitbitHelper.getLastSyncedTimestamp(accessToken, fitbitUserId);
-		const ranges = FitbitHelper.getTimeRanges(startTimestamp, endTimestamp);
-
-		console.log("RateLimit: "+RateLimit.totalQuota)
-		console.log("Refill: "+RateLimit.remainingSecondsUntilRefill)
-		console.log("Number of request remainig in processing "+ requestType + " : " + ranges.length);
-		for (const range of ranges) {
-			if(RateLimit.isLimitExceeded()) {
-				console.log("Request limit is exceeded!");
-				RateLimit.setProcessedStatus(userId, false);
-				break;
-			}
-			let response = await ApiManager.getIntradayByDateAndTime(accessToken, fitbitUserId, requestType, range, detailLevel);
-			await DBManager.storeIntradayData(userId, requestType, range, response);
-
-			RateLimit.requestProcessed();
-		}
-		console.log("Total " + RateLimit.numberOfRequestProcessed + " Request processed successfully!");
-	}
-
-	static async processIntradayByInterval(userId, accessToken, fitbitUserId, requestType, maximumRange) {
-		console.log("Processing "+ requestType);
-		let startTimestamp = await FitbitHelper.getLastPolledTimestamp(userId, requestType);
-		let endTimestamp = await FitbitHelper.getLastSyncedTimestamp(accessToken, fitbitUserId);
-		const ranges = FitbitHelper.getDateAndTimeRanges(startTimestamp, endTimestamp, maximumRange);
-
-		console.log("RateLimit: "+RateLimit.totalQuota)
-		console.log("Refill: "+RateLimit.remainingSecondsUntilRefill)
-		console.log("Number of request remainig in processing "+ requestType + " : " + ranges.length);
-		for (const range of ranges) {
-			if(RateLimit.isLimitExceeded()) {
-				console.log("Request limit is exceeded!");
-				RateLimit.setProcessedStatus(userId, false);
-				break;
-			}
-			let response = await ApiManager.getIntradayByInterval(accessToken, fitbitUserId, requestType, range);
-			await DBManager.storeIntradayByIntervalData(userId, requestType, range, response);
-
-			RateLimit.requestProcessed();
-		}
-		console.log("Total " + RateLimit.numberOfRequestProcessed + " Request processed successfully!");
-	}
-
-	static async processTimeSeriesByDateRange(userId, accessToken, fitbitUserId, requestType, maximumRange) {
-		console.log("Processing "+ requestType);
-		let startTimestamp = await FitbitHelper.getLastPolledTimestamp(userId, requestType);
-		let endTimestamp = await FitbitHelper.getLastSyncedTimestamp(accessToken, fitbitUserId);
-		const ranges = FitbitHelper.getDateAndTimeRanges(startTimestamp, endTimestamp, maximumRange);
-
-		console.log("RateLimit: "+RateLimit.totalQuota)
-		console.log("Refill: "+RateLimit.remainingSecondsUntilRefill)
-		console.log("Number of request remainig in processing "+ requestType + " : " + ranges.length);
-		for (const range of ranges) {
-			if(RateLimit.isLimitExceeded()) {
-				console.log("Request limit is exceeded!");
-				RateLimit.setProcessedStatus(userId, false);
-				break;
-			}
-			let response = await ApiManager.getTimeSeriesByDateRange(accessToken, fitbitUserId, requestType, range);
-			await DBManager.storeTimeSeriesByDateRange(userId, requestType, range, response);
-
-			RateLimit.requestProcessed();
-		}
-		console.log("Total " + RateLimit.numberOfRequestProcessed + " Request processed successfully!");
 	}
 }
 module.exports = FitbitManager;
